+++ src/project/rollup-plugin-gro-json.ts
@@ -1,7 +1,7 @@
 import {Plugin} from 'rollup';
 import {createFilter, dataToEsm} from '@rollup/pluginutils';
 
-import {magenta} from '../colors/terminal.js';
+import {magenta} from '../utils/terminal.js';
 import {SystemLogger} from '../utils/log.js';
 import {printPath} from '../utils/print.js';
 import {omitUndefined} from '../utils/object.js';