+++ src/project/rollup-plugin-gro-typescript.ts
@@ -3,7 +3,7 @@ import {Plugin, PluginContext} from 'rollup';
 import {resolve} from 'path';
 import {createFilter} from '@rollup/pluginutils';
 
-import {magenta, red} from '../colors/terminal.js';
+import {magenta, red} from '../utils/terminal.js';
 import {createStopwatch} from '../utils/time.js';
 import {SystemLogger, Logger} from '../utils/log.js';
 import {printKeyValue, printMs, printPath} from '../utils/print.js';